refactor: 불필요한 코드 제거, 컨벤션 적용
