+++ src/main/java/woowacourse/shoppingcart/application/CustomerService.java
@@ -1,8 +1,8 @@
 package woowacourse.shoppingcart.application;
 
 import org.springframework.stereotype.Service;
-import woowacourse.auth.domain.Customer;
 import woowacourse.shoppingcart.dao.CustomerDao;
+import woowacourse.shoppingcart.domain.Customer;
 import woowacourse.shoppingcart.dto.CustomerResponse;
 import woowacourse.shoppingcart.dto.DeleteCustomerRequest;
 import woowacourse.shoppingcart.dto.SignUpRequest;
@@ -41,12 +41,12 @@ public class CustomerService {
     }
 
     public void deleteMe(String username, DeleteCustomerRequest deleteCustomerRequest) {
-        if (!customerDao.existByUserName(username)) {
+        if (!customerDao.existByUsername(username)) {
             throw new InvalidCustomerException();
         }
         if (!customerDao.isValidPasswordByUsername(username, deleteCustomerRequest.getPassword())) {
             throw new InvalidPasswordException();
         }
-        customerDao.deleteByUserName(username);
+        customerDao.deleteByUsername(username);
     }
 }
