+++ src/main/java/woowacourse/shoppingcart/dao/CustomerDao.java
@@ -9,7 +9,7 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import woowacourse.auth.domain.Customer;
+import woowacourse.shoppingcart.domain.Customer;
 import woowacourse.shoppingcart.exception.InvalidCustomerException;
 
 @Repository
@@ -18,7 +18,7 @@ public class CustomerDao {
     private final JdbcTemplate jdbcTemplate;
     private final SimpleJdbcInsert simpleJdbcInsert;
 
-    private RowMapper<Customer> customerRowMapper = (rs, rowNum) -> new Customer(
+    private final RowMapper<Customer> customerRowMapper = (rs, rowNum) -> new Customer(
             rs.getLong("id"),
             rs.getString("username"),
             rs.getString("email"),
@@ -55,7 +55,7 @@ public class CustomerDao {
         }
     }
 
-    public boolean existByUserName(String username) {
+    public boolean existByUsername(String username) {
         try {
             String query = "SELECT EXISTS (SELECT * FROM customer WHERE username = ?)";
             return jdbcTemplate.queryForObject(query, Boolean.class, username);
@@ -87,7 +87,7 @@ public class CustomerDao {
         jdbcTemplate.update(query, password, id);
     }
 
-    public void deleteByUserName(String username) {
+    public void deleteByUsername(String username) {
         final String query = "DELETE FROM customer WHERE username = ?";
         jdbcTemplate.update(query, username);
     }
